DEV-0000/1/1: refactored create
