+++ src/services/Path.js
@@ -1,0 +1,1 @@
+export const API = 'http://localhost:5000/api/items';
